refactor: deduplicate copy-paste code into shared helpers/mixins (#495)
Collapse genuine copy-paste duplication into shared helpers/mixins with no
behavior change; existing tests cover all sites.

- check_pointer: call existing _matches_discriminator instead of an inline
  reimplementation of the same options-matching loop.
- BaseInputData: lift the identical load() body into the parent; declare
  init_reader on the base; drop the three subclass copies.
- TimeReferenceMixin: extract the verbatim get_reference_time_column and
  TIME_UNITS shared by TimeWindowFeatureGroup and ForecastingFeatureGroup.
- FeatureChainParserMixin._extract_operation_and_source_feature: generic
  helper backing six _extract_*_and_source_feature methods (error text
  preserved via a label argument).
- BaseFilterEngine.do_max_filter: move the validation/branching skeleton to
  the base; each backend overrides two one-line leaf helpers
  (_apply_max_exclusive_filter / _apply_max_inclusive_filter).
