+++ mloda_plugins/compute_framework/base_implementations/pandas/pandas_filter_engine.py
@@ -29,37 +29,12 @@ class PandasFilterEngine(BaseFilterEngine):
         return data[data[filter_feature.name] >= value]
 
     @classmethod
-    def do_max_filter(cls, data: Any, filter_feature: SingleFilter) -> Any:
-        # Check if this is a complex parameter with max/max_exclusive or a simple one with value
-        has_max = filter_feature.parameter.max_value is not None
-        has_value = filter_feature.parameter.value is not None
+    def _apply_max_exclusive_filter(cls, data: Any, column_name: str, threshold: Any) -> Any:
+        return data[data[column_name] < threshold]
 
-        if has_max:
-            # Complex parameter - use get_min_max_operator
-            min_parameter, max_parameter, max_operator = cls.get_min_max_operator(filter_feature)
-
-            if min_parameter is not None:
-                raise ValueError(
-                    f"Filter parameter {filter_feature.parameter} not supported as max filter: {filter_feature.name}"
-                )
-
-            if max_parameter is None:
-                raise ValueError(
-                    f"Filter parameter {filter_feature.parameter} is None although expected: {filter_feature.name}"
-                )
-
-            if max_operator is True:
-                return data[data[filter_feature.name] < max_parameter]
-            else:
-                return data[data[filter_feature.name] <= max_parameter]
-        elif has_value:
-            # Simple parameter - extract the value
-            value = filter_feature.parameter.value
-            if value is None:
-                raise ValueError(f"Filter parameter 'value' not found in {filter_feature.parameter}")
-            return data[data[filter_feature.name] <= value]
-        else:
-            raise ValueError(f"No valid filter parameter found in {filter_feature.parameter}")
+    @classmethod
+    def _apply_max_inclusive_filter(cls, data: Any, column_name: str, threshold: Any) -> Any:
+        return data[data[column_name] <= threshold]
 
     @classmethod
     def do_equal_filter(cls, data: Any, filter_feature: SingleFilter) -> Any:
