+++ mloda_plugins/feature_group/experimental/time_window/base.py
@@ -17,9 +17,10 @@ from mloda.user import FeatureName
 from mloda.provider import FeatureSet
 from mloda.user import Options
 from mloda.provider import DefaultOptionKeys
+from mloda_plugins.feature_group.experimental.time_reference_mixin import TimeReferenceMixin
 
 
-class TimeWindowFeatureGroup(FeatureChainParserMixin, FeatureGroup):
+class TimeWindowFeatureGroup(TimeReferenceMixin, FeatureChainParserMixin, FeatureGroup):
     """
     Base class for all time window feature groups.
 
@@ -75,27 +76,6 @@ class TimeWindowFeatureGroup(FeatureChainParserMixin, FeatureGroup):
     WINDOW_SIZE = "window_size"
     TIME_UNIT = "time_unit"
 
-    @classmethod
-    def get_reference_time_column(cls, options: Optional[Options] = None) -> str:
-        """
-        Get the reference time column name from options or use the default.
-
-        Args:
-            options: Optional Options object that may contain a custom reference time column name
-
-        Returns:
-            The reference time column name to use
-        """
-        reference_time_key = DefaultOptionKeys.reference_time
-        if options and options.get(reference_time_key):
-            reference_time = options.get(reference_time_key)
-            if not isinstance(reference_time, str):
-                raise ValueError(
-                    f"Invalid reference_time option: {reference_time}. Must be string. Is: {type(reference_time)}."
-                )
-            return reference_time
-        return DefaultOptionKeys.reference_time
-
     # Define supported window functions
     WINDOW_FUNCTIONS = {
         "sum": "Sum of values in window",
@@ -109,17 +89,6 @@ class TimeWindowFeatureGroup(FeatureChainParserMixin, FeatureGroup):
         "median": "Median value in window",
         "first": "First value in window",
         "last": "Last value in window",
-    }
-
-    # Define supported time units
-    TIME_UNITS = {
-        "second": "Seconds",
-        "minute": "Minutes",
-        "hour": "Hours",
-        "day": "Days",
-        "week": "Weeks",
-        "month": "Months",
-        "year": "Years",
     }
 
     # Define PROPERTY_MAPPING for the new unified parser approach
@@ -141,7 +110,7 @@ class TimeWindowFeatureGroup(FeatureChainParserMixin, FeatureGroup):
         },
         # Time unit parameter (context parameter)
         TIME_UNIT: {
-            **TIME_UNITS,  # Reference existing TIME_UNITS dict
+            **TimeReferenceMixin.TIME_UNITS,  # Reference shared TIME_UNITS dict
             DefaultOptionKeys.context: True,  # Mark as context parameter
             DefaultOptionKeys.strict_validation: True,  # Enable strict validation
         },
