+++ mloda_plugins/feature_group/input_data/read_file.py
@@ -66,23 +66,6 @@ class ReadFile(BaseInputData):
     def get_column_names(cls, file_name: str) -> list[str]:
         raise NotImplementedError
 
-    def load(self, features: FeatureSet) -> Any:
-        _options = None
-
-        for feature in features.features:
-            if _options:
-                if _options != feature.options:
-                    raise ValueError("All features must have the same options.")
-            _options = feature.options
-
-        reader, data_access = self.init_reader(_options)
-        data = reader.load_data(data_access, features)
-
-        if data is None:
-            raise ValueError(f"Loading data failed for feature {features.get_name_of_one_feature()}.")
-
-        return data
-
     def init_reader(self, options: Optional[Options]) -> tuple["ReadFile", Any]:
         if options is None:
             raise ValueError(
